third_party/blink/web_tests/platform/generic/http/tests/devtools/profiler/heap-snapshot-containment-show-all-expected.txt

Tests Containment view of detailed heap snapshots. The "Show All" button must show all nodes. Test object distances calculation.

Profiler was enabled.
Detailed heap profiles were enabled.

Running: testShowAll

Running: testDistances

Profiler was disabled.
